[feat] 로그인/회원가입 컴포넌트 통합 및 회원가입 마크업
